// bamul/src/app/shared/model/vehicle.ts
export class Vehicle {
    vehicleRegNumber: string;
    modelYear: number; /*pending year*/
    make: string;
    model: string;
    registrationState: string;/*pending year*/
    registrationDate: Date;
    regPeriodTo: Date;
    regPeriodFrom: Date;
    vehicleType: string;
    vehicleSubType: string;
    materialType: string;
    capacity: number;
    capacityUnits: string;
    status: string;
    fuelType: string;
    recordCreationBy: any;
    recordCreationDate: any;
    lastUpdateDate: any;
    lastUpdateBy: any;
    vehicleId: number;
}